requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
cryptography==43.0.1
cssselect==1.2.0
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
discord==2.3.2
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
fake-useragent==1.5.1
fastapi==0.109.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.9.0
fugle-trade==1.2.0
fugle-trade-core==2.1.0
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
importlib_metadata==8.4.0
iniconfig==2.0.0
jeepney==0.8.0
jieba3k==0.35.1
jiter==0.5.0
joblib==1.4.2
keyring==23.5.0
keyrings.cryptfile==1.3.8
llama-cloud==0.0.17
llama-deploy==0.1.0
llama-index==0.11.7
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.0
llama-index-core==0.11.7
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-ollama==0.3.1
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.1
llama-index-readers-llama-parse==0.3.0
llama-index-readers-web==0.2.2
llama-index-tools-google==0.2.0
llama-index-tools-tavily-research==0.2.0
llama-index-tools-yahoo-finance==0.2.0
llama-parse==0.5.2
lxml==5.3.0
marshmallow==3.22.0
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.2
openai==1.44.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parse==1.20.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
playwright==1.47.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
pyee==12.0.0
pyparsing==3.1.4
pypdf==4.3.1
pyquery==2.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
requests-html==0.10.0
requests-oauthlib==2.0.0
rsa==4.9
SecretStorage==3.3.3
selenium==4.24.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.0.27
SQLAlchemy==2.0.34
starlette==0.36.3
striprtf==0.0.26
tavily-python==0.5.0
tenacity==8.5.0
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.2
tomli==2.0.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.30.6
w3lib==2.2.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.10.0
yfinance==0.2.43
zipp==3.20.1
